now can do actions based on speech recognition

+++ say.sh
@@ -4,28 +4,22 @@
 # Logitech HD Webcam C270, 720p Widescreen Video Calling and Recording
 
 
-echo "Recording... Press Ctrl+C to Stop."
-rm file.wav
-rm file.flac
-#arecord -D "plughw:1,0" -d4 -f dat -t wav -r 48000 -c 2 | sox - -b16 -r16k -c1 -t wav - | flac - -o file.flac
+duration=$1
 
-espeak "speak now"
-arecord -f S16_LE -d 2 file.wav  -D "hw:0,0"  -r 16000 -c2
-espeak "stop it"
+echo "Recording... Press Ctrl+C to Stop."
+if [ -f file.wav ] ; then
+rm file.wav
+fi
+
+if [ -f file.flac ] ; then
+rm file.flac
+fi
+
+arecord -f S16_LE -d $duration file.wav  -D "hw:0,0"  -r 16000 -c2
+
 
 ffmpeg -loglevel panic -y -i file.wav -ar 16000 -acodec flac file.flac  > /dev/null 2>&1
-#arecord -d 2 -D "plughw:1,0" -q -f cd -t wav file.wav
-#echo "processing 1"
-#arecord -D "plughw:1,0" --duration=3 -f cd --file-type wav file.wav
-#| ffmpeg -loglevel panic -y -i -ar 16000 -acodec flac file.flac  > /dev/null 2>&1
-#flac --best -f file.wav
 
-
-#wget --post-file='file.flac' \
-#--user-agent='Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_8) AppleWebKit/535.7 (KHTML, like Gecko) Chrome/16.0.912.77 Safari/535.7' \
-#--header='Content-Type: audio/x-flac; rate=16000;' \
-#-O 'recognized.json' \
-#'https://www.google.com/speech-api/v1/recognize?client=chromium&lang=ar-QA&maxresults=10'
 
 echo "Processing..."
 curl -X POST \
@@ -34,11 +28,5 @@ curl -X POST \
 --header 'Content-Type: audio/x-flac; rate=16000;' \
 'https://www.google.com/speech-api/v2/recognize?output=json&lang=en-us&key=AIzaSyCnl6MRydhw_5fLXIdASxkLJzcJh5iX0M4' > test.txt
 
-cat test.txt
-
-
-#curl -X POST  --data-binary @file.flac  --user-agent 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_8) AppleWebKit/535.7 (KHTML, like Gecko) Chrome/16.0.912.77 Safari/535.7'  --header 'Content-Type: audio/x-flac; rate=16000;'  'https://www.google.com/speech-api/v1/recognize?client=chromium&lang=ar-QA&maxresults=10'
-#
-#
-# wget -q -U "Mozilla/5.0" --post-file file.flac --header "Content-Type: audio/x-flac; rate=16000" -O - "http://www.google.com/speech-api/v1/recognize?lang=en-us&client=chromium" | cut -d\" -f12  >stt.txt
-#
+response=`cat test.txt | grep transcript | jq '.result[0].alternative[0].transcript'`
+espeak "you said $response"
